Añadido de mas controles y mensajes

+++ Frontend/src/app/components/carrito/carrito.component.ts
@@ -3,6 +3,9 @@ import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 import { UsuarioService } from '../../services/usuarios.service';
 import { FormBuilder, FormGroup, ReactiveFormsModule, Validators } from '@angular/forms';
+import Swal from 'sweetalert2';
+import { lastValueFrom } from 'rxjs';
+
 
 @Component({
   selector: 'app-carrito',
@@ -24,7 +27,7 @@ carrito: any[] = [];
 
   constructor(private fb: FormBuilder,private router: Router, private usuarioService: UsuarioService) {
    this.userForm = this.fb.group({
-      telefono: ['', [Validators.required,Validators.pattern(/^\+?[0-9]{7,15}$/)]],
+      telefono: ['', [Validators.required,Validators.pattern(/^\+?[0-9]{4,15}$/)]],
       departamento: ['', Validators.required],
       direccion: ['', Validators.required]
     });
@@ -77,71 +80,47 @@ verProducto(id : string) {
 
   comprar() {
     this.quiereComprar = true;
-    return;
-    this.usuarioService.crearCompra(this.user,this.total).subscribe({
-      
-    });
-
-    for (let i = 0; i < this.carrito.length; i++) {
-      
-    this.usuarioService.comprarCarrito(this.user, this.carrito[i].id,this.carrito[i].precio,this.carrito[i].cantidad).subscribe({
-        
-});
-
-  }
-  this.eliminarTodos();
-this.carrito.length = 0;
-
-
-      if(this.carrito.length === 0) {
-      this.hayProductos = false;
-      }
 }
 
 get f() {
     return this.userForm.controls;
   }
 
-onSubmit() {
-    this.submitted = true;
-        console.log(this.userForm.get('telefono')?.value);
-                console.log(this.userForm.get('departamento')?.value);
+async onSubmit() {
+ 
+  this.submitted = true;
+  
+   if (this.userForm.invalid) return;
 
-                        console.log(this.userForm.get('direccion')?.value);
-
-
-    if (this.userForm.invalid) return;
-    return;
-
-    this.usuarioService.crearCompra(this.user,this.total).subscribe({
-      
-    });
-
+    let telefonoNumber = this.userForm.get('telefono')?.value;
+    console.log(telefonoNumber);
+    try{
+    await lastValueFrom( this.usuarioService.crearCompra(this.user,this.total - 200,telefonoNumber,this.userForm.get('direccion')?.value,this.userForm.get('departamento')?.value));
     for (let i = 0; i < this.carrito.length; i++) {
       
-    this.usuarioService.comprarCarrito(this.user, this.carrito[i].id,this.carrito[i].precio,this.carrito[i].cantidad).subscribe({
-        
-});
+    await lastValueFrom (this.usuarioService.comprarCarrito(this.user, this.carrito[i].id,this.carrito[i].precio,this.carrito[i].cantidad));
 
   }
-  this.eliminarTodos();
-this.carrito.length = 0;
-this.userForm.reset();
-        this.submitted = false;
-
-      if(this.carrito.length === 0) {
-      this.hayProductos = false;
+  
+  Swal.fire({
+          icon: 'success',
+          title: '¡Compra realizada con éxito!',
+          showConfirmButton: false,
+          timer: 2000
+        }).then(() => {
+          
+          this.eliminarTodos();
+          this.carrito.length = 0;
+          this.userForm.reset();
+          this.submitted = false;
+          this.hayProductos = false;
+        });
+      }catch(error){
+        console.error("Error en la compra",error);
       }
+
+       
   }
-
-
-
-
-
-
-
-
-
 
 eliminarTodos(){
   for (let i = 0; i < this.carrito.length; i++) {
@@ -155,4 +134,6 @@ eliminarTodos(){
     });
   }
 }
+
+
 }
